# docs/swagger.yaml
basePath: /
definitions:
  models.Category:
    properties:
      id_category:
        type: integer
      name:
        type: string
    type: object
  models.Meme:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      id_meme:
        type: integer
      id_user:
        type: integer
      image:
        type: string
      long_description:
        type: string
      name:
        type: string
      short_description:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      user:
        $ref: '#/definitions/models.User'
      vector_list:
        items:
          type: number
        type: array
    type: object
  models.Tag:
    properties:
      id_tag:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      id_role:
        type: integer
      id_user:
        type: integer
      login:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  models.UserRole:
    properties:
      id_role:
        type: integer
      name_role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Система пояснения сложных информационных сообщений юмористического
    характера.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MemesApp
  version: "0.2"
paths:
  /categories:
    patch:
      consumes:
      - application/json
      description: Обновляет категорию мемов по указанному ID
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      - description: Обновляемые поля
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить категорию мемов
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Добавляет новую категорию мемов в базу данных
      parameters:
      - description: Данные новой категории
        in: body
        name: meme
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новую категорию мемов
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет категорию мемов по указанному ID
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить категорию мемов
      tags:
      - Categories
  /memes:
    get:
      consumes:
      - application/json
      description: Возвращает список всех мемов из базы данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Meme'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить все мемы
      tags:
      - Memes
    post:
      consumes:
      - application/json
      description: Добавляет новый мем в базу данных
      parameters:
      - description: Данные нового мема
        in: body
        name: meme
        required: true
        schema:
          $ref: '#/definitions/models.Meme'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Meme'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новый мем
      tags:
      - Memes
  /memes/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет мем по указанному ID
      parameters:
      - description: ID мема
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить мем
      tags:
      - Memes
    get:
      consumes:
      - application/json
      description: Возвращает мем с указанным ID
      parameters:
      - description: ID мема
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meme'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить мем по ID
      tags:
      - Memes
    patch:
      consumes:
      - application/json
      description: Обновляет данные мема по ID
      parameters:
      - description: ID мема
        in: path
        name: id
        required: true
        type: integer
      - description: Обновляемые поля
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить мем
      tags:
      - Memes
  /tags:
    post:
      consumes:
      - application/json
      description: Добавляет новый тег в базу данных
      parameters:
      - description: Данные нового тега
        in: body
        name: meme
        required: true
        schema:
          $ref: '#/definitions/models.Tag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Tag'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новый тег
      tags:
      - Tags
  /tags/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет тег по указанному ID
      parameters:
      - description: ID тега
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить тег
      tags:
      - Tags
    patch:
      consumes:
      - application/json
      description: Обновляет данные тега по ID
      parameters:
      - description: ID тега
        in: path
        name: id
        required: true
        type: integer
      - description: Обновляемые поля
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновть тег
      tags:
      - Tags
swagger: "2.0"